refactor: 게임결과 클래스 패키지 변경
