+++ src/test/java/nextstep/ladder/domain/ladder/ResultsTest.java
@@ -1,5 +1,6 @@
  package nextstep.ladder.domain.ladder;
 
+import nextstep.ladder.domain.player.Results;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
